+++ cilantro/include/graphics/Framebuffer.h
@@ -6,31 +6,31 @@
 class CFramebuffer : public IFramebuffer
 {
 public:
-    CFramebuffer (uint32_t bufferWidth, uint32_t bufferHeight, size_t rgbTextureCount, size_t rgbaTextureCount, size_t depthBufferArrayLayerCount, bool depthStencilRenderbufferEnabled);
+    CFramebuffer (unsigned int bufferWidth, unsigned int bufferHeight, unsigned int rgbTextureCount, unsigned int rgbaTextureCount, unsigned int depthBufferArrayLayerCount, bool depthStencilRenderbufferEnabled);
     virtual ~CFramebuffer () {};
 
     ///////////////////////////////////////////////////////////////////////////
 
-    virtual uint32_t GetWidth () const override final;
-    virtual uint32_t GetHeight () const override final;
+    virtual unsigned int GetWidth () const override final;
+    virtual unsigned int GetHeight () const override final;
 
-    virtual size_t GetColorTextureCount() const override final;
-    virtual size_t GetRGBTextureCount () const override final;
-    virtual size_t GetRGBATextureCount () const override final;
-    virtual size_t GetDepthArrayLayerCount () const override final;
+    virtual unsigned int GetColorTextureCount() const override final;
+    virtual unsigned int GetRGBTextureCount () const override final;
+    virtual unsigned int GetRGBATextureCount () const override final;
+    virtual unsigned int GetDepthArrayLayerCount () const override final;
 
     virtual bool IsDepthStencilRenderbufferEnabled () const override final;
     virtual bool IsDepthArrayEnabled () const override final;
 
-    void SetFramebufferResolution (uint32_t bufferWidth, uint32_t bufferHeight) override;
+    void SetFramebufferResolution (unsigned int bufferWidth, unsigned int bufferHeight) override;
 
     ///////////////////////////////////////////////////////////////////////////
 
 protected:
-    size_t      m_rgbTextureCount;
-    size_t      m_rgbaTextureCount;
-    size_t      m_depthBufferArrayLayerCount;
-    uint32_t    m_bufferWidth;
-    uint32_t    m_bufferHeight;
-    bool        m_depthStencilRenderbufferEnabled;
+    unsigned int    m_rgbTextureCount;
+    unsigned int    m_rgbaTextureCount;
+    unsigned int    m_depthBufferArrayLayerCount;
+    unsigned int    m_bufferWidth;
+    unsigned int    m_bufferHeight;
+    bool            m_depthStencilRenderbufferEnabled;
 };
